Renamed back some of the wasm files
